(PC-23087)[PRO] refactor: clean icon stroke-trash
